Large commit, with multiple usability improvements and bugfixes:
1) In case of OpenStack clouds, CBTOOL can now dynamically create
tenant/user/net/subnet/router before attaching a new VM or VApp. This is done
be using the generic script execution capabilities in CBTOOL, which can be
configured to become active at multiple steps during attachment.

    Here is how to use it in order to have each VM on its own tenants:

    a) First configure the "vm_defaults" object for the whole experiment:
    cldalter vm_defaults execute_script_name=~/cloudbench/scenarios/scripts/openstack_multitenant.sh

    b) Attach new VMs by adding the parameter "execute_provision_originated"
    vmattach tinyvm auto empty default execute_provision_originated

    Here is how to use it in order to have each AI (VApp) on its own tenants:

    a) First configure the "ai_defaults" object for the whole experiment:

    cldalter ai_defaults execute_script_name=~/cloudbench/scenarios/scripts/openstack_multitenant.sh

    b) Attach new AIs by adding the parameter "execute_provision_originated"
aiattach nullworkload default default none none execute_provision_originated

    This script will also attempt to collect data from StackTach - if avaible -
    by using the "stacky" command line, and will make sure that the
    tenant/user/net/subnet/router is removed when the VM is destroyed.

2) Add two new commands specially for debugging purposes. The first, "vmdev", will perform VM attachments
 and then display (instead of executing) all the commands issued by the orchestrator.
 This debugging mode can be disabled by issuing "vmundev". The second command
"appdev", operates in a similar manner, displaying all the commands that the
orchestrator would execute to fully deploy a Virtual Application (similarly, "appdundev" will disable it).

3) Added a Docker file to build a fully self-contained orchestrator node. Build
 it by executing "cd /cbtool/docker/orchestrator/; docker build ." and the run it
 with "docker run -it --privileged=true cb_orchestrator"

4) Bugfixes for Google Compute Engine (GCE): made sure that multi-threaded
instance creation can handle the VApps with the size required by SPECCLoud, a
sanitized hostnames to avoid problematic characters, such as colon.

5) Bugfixes for OpenStack: wait until the instance is fully started before
attempting to attach the floating IP, make sure that the floating IPs are deleted
 upon VM destruction, and updated novaclient to use the versioning mechanism.

6) Multiple bugfixes for different workloads. Workload functionality is now
checked by cbtool/clients/real_application_regression_test.py

+++ scripts/common/cb_start_load_manager.sh
@@ -48,6 +48,16 @@ then
 fi
 
 load_manager_vm=`get_my_ai_attribute load_manager_vm`
+run_application_scripts=`get_my_ai_attribute run_application_scripts`
+run_application_scripts=`echo $run_application_scripts | tr '[:upper:]' '[:lower:]'`
+
+debug_remote_commands=`get_my_ai_attribute debug_remote_commands`
+debug_remote_commands=`echo $debug_remote_commands | tr '[:upper:]' '[:lower:]'`
+
+if [[ $run_application_scripts == "false" && $debug_remote_commands == "true" ]]
+then
+    daemonize="--logdest console -v 5"
+fi
 
 cat > /tmp/cbloadman <<EOF
 #!/usr/bin/env bash
@@ -67,23 +77,29 @@ sudo chmod 755 /usr/local/bin/cbloadman
 
 if [[ x"${load_manager_vm}" == x"${my_vm_uuid}" ]]
 then
-    running_load_managers=`pgrep -f $operation`
-
-    if [[ x"${running_load_managers}" == x ]]
+    if [[ $run_application_scripts == "false" && $debug_remote_commands == "true" ]]
     then
-        syslog_netcat "Starting Load Manager"
-        /usr/local/bin/cbloadman
+        syslog_netcat "The command \"appdev\" was run on the orchestrator. Please start the Load Manager in debug mode by running /usr/loca/bin/cbloadman"    
+    else        
+        running_load_managers=`pgrep -f $operation`
+    
+        if [[ x"${running_load_managers}" == x ]]
+        then
+            syslog_netcat "Starting Load Manager"
+            /usr/local/bin/cbloadman
 
-	    sudo pkill -9 -f cbloadmanwatch
-		sudo screen -d -m -S cbloadmanwatch bash -c "su - ${my_login_username}"
-		sudo screen -p 0 -S cbloadmanwatch -X stuff "while true ; do sleep 30; /usr/local/bin/cbloadman; done$(printf \\r)"				    	
-
-        exit 0
-    else
-        syslog_netcat "A Load Manager is already running"
-        exit 2
+            sudo pkill -9 -f cbloadmanwatch
+            sudo screen -d -m -S cbloadmanwatch bash -c "su - ${my_login_username}"
+            sudo screen -p 0 -S cbloadmanwatch -X stuff "while true ; do sleep 30; /usr/local/bin/cbloadman; done$(printf \\r)"                        
+    
+            exit 0
+        else
+            syslog_netcat "A Load Manager is already running"
+            exit 2
+        fi
     fi
 fi
 syslog_netcat "This VM is not designated as Load Manager"
+
 exit 0
 exit 0
